Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -2,7 +2,7 @@
 int main(){
     int a,b;
     char c;
-    float division ;
+    int division ;
     division = a/b;
     scanf("%d %d %c",&a,&b,&c);
     if(c=='+'){
